fix showing last page in onboardiing
