+++ app/src/main/java/com/barcelona/qurio/di/NetworkModule.kt
@@ -1,9 +1,5 @@
 package com.barcelona.qurio.di
 
-import android.content.Context
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.PreferenceDataStoreFactory
-import androidx.datastore.preferences.core.Preferences
 import com.barcelona.qurio.BuildConfig
 import com.barcelona.qurio.model.api.TriviaApiService
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
