+++ app/src/main/java/com/barcelona/qurio/service/UserPreferences.kt
@@ -1,0 +1,8 @@
+package com.barcelona.qurio.service
+
+import kotlinx.coroutines.flow.Flow
+
+interface UserPreferences {
+    val isFirstLaunch: Flow<Boolean>
+    suspend fun setFirstLaunch()
+}
